+++ .evergreen/run-tests.sh
@@ -78,12 +78,12 @@ elif [ "$COMPRESSORS" = "snappy" ]; then
     createvirtualenv $PYTHON_BINARY snappytest
     trap "deactivate; rm -rf snappytest" EXIT HUP
     # 0.5.2 has issues in pypy3(.5)
-    pip install python-snappy==0.5.1
+    python -m pip install python-snappy==0.5.1
     PYTHON=python
 elif [ "$COMPRESSORS" = "zstd" ]; then
     createvirtualenv $PYTHON_BINARY zstdtest
     trap "deactivate; rm -rf zstdtest" EXIT HUP
-    pip install zstandard
+    python -m pip install zstandard
     PYTHON=python
 else
     PYTHON="$PYTHON_BINARY"